388 adding selection of examples when creating a test
