+++ web/src/constants/Test.constants.ts
@@ -1,0 +1,40 @@
+import {HTTP_METHOD} from './Common.constants';
+
+interface DemoTestExample {
+  name: string;
+  url: string;
+  method: HTTP_METHOD;
+  body: string;
+  description: string;
+}
+
+export const DemoTestExampleList: DemoTestExample[] = [
+  {
+    name: 'Shopping app',
+    url: 'http://shop/buy',
+    method: HTTP_METHOD.GET,
+    body: '',
+    description: 'Generic get',
+  },
+  {
+    name: 'Pokemon - List',
+    url: 'http://demo-pokemon-api.demo.svc.cluster.local/pokemon?take=20&skip=0',
+    method: HTTP_METHOD.GET,
+    body: '',
+    description: 'Get a Pokemon',
+  },
+  {
+    name: 'Pokemon - Add',
+    url: 'http://demo-pokemon-api.demo.svc.cluster.local/pokemon',
+    method: HTTP_METHOD.POST,
+    body: '{"name":"meowth","type":"normal","imageUrl":"https://assets.pokemon.com/assets/cms2/img/pokedex/full/052.png","isFeatured":true}',
+    description: 'Add a Pokemon',
+  },
+  {
+    name: 'Pokemon - Import',
+    url: 'http://demo-pokemon-api.demo.svc.cluster.local/pokemon/import',
+    method: HTTP_METHOD.POST,
+    body: '{"id":52}',
+    description: 'Import a Pokemon',
+  },
+];
