+++ web/src/models/Trace.model.ts
@@ -1,7 +1,7 @@
 import {IRawTrace, ITrace} from '../types/Trace.types';
 import Span from './Span.model';
 
-const Trace = ({description, resourceSpans}: IRawTrace): ITrace => {
+const Trace = ({description = '', resourceSpans = []}: IRawTrace): ITrace => {
   return {
     description,
     spans: Span.createFromResourceSpanList(resourceSpans),
