+++ web/src/types/Trace.types.ts
@@ -1,8 +1,8 @@
 import {IResourceSpan, ISpan} from './Span.types';
 
 export type IRawTrace = {
-  resourceSpans: Array<IResourceSpan>;
-  description: string;
+  resourceSpans?: Array<IResourceSpan>;
+  description?: string;
 };
 
 export type ITrace = {
